refactor(protocol): abstract status logging into own module
